refactor : styles index 수정
