+++ src/app/styles/device.ts
@@ -3,4 +3,5 @@ export const device = {
   TABLET: '@media (min-width: 37.5rem) and (max-width: 76.5rem)',
   DESKTOP: '@media (min-width: 76.5rem)',
 };
+
 export type DeviceType = typeof device;
