+++ src/app/styles/globalStyle.ts
@@ -4,13 +4,12 @@ import reset from 'styled-reset';
 import Pretendard from '@/shared/assets/font/PretendardVariable.woff2';
 
 export const GlobalStyle = createGlobalStyle`
-
+  ${reset}
   @font-face {
     font-family: 'Pretendard';
     src: local('Pretendard'), url(${Pretendard}) format('woff2');
     font-weight: 100 900;
   }
-  
   html{
     font-size:62.5%;
   }
