+++ src/app/styles/index.ts
@@ -1,2 +1,8 @@
 export { GlobalStyle } from '@/app/styles/globalStyle';
 export { theme } from '@/app/styles/theme';
+export { color } from './color';
+export { device } from './device';
+export { typo } from './typo';
+export type { ColorType } from '@/app/styles/color';
+export type { TypoType } from '@/app/styles/typo';
+export type { DeviceType } from '@/app/styles/device';
